+++ edukit-core/src/main/java/com/edukit/core/studentrecord/metric/StudentRecordMetricsAspect.java
@@ -1,9 +1,7 @@
-package com.edukit.core.studentrecord.aop;
+package com.edukit.core.studentrecord.metric;
 
 import com.edukit.core.studentrecord.db.entity.StudentRecord;
 import com.edukit.core.studentrecord.db.enums.StudentRecordType;
-import com.edukit.core.studentrecord.service.RecordGenerationTracker;
-import com.edukit.core.studentrecord.service.StudentRecordMetricsCounter;
 import com.edukit.core.studentrecord.service.StudentRecordService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
