+++ edukit-core/src/main/java/com/edukit/core/studentrecord/metric/StudentRecordMetricsCounter.java
@@ -1,4 +1,4 @@
-package com.edukit.core.studentrecord.service;
+package com.edukit.core.studentrecord.metric;
 
 import com.edukit.core.studentrecord.db.enums.StudentRecordType;
 import io.micrometer.core.instrument.MeterRegistry;
